+++ app/utils/WkConf.scala
@@ -89,6 +89,7 @@ class WkConf @Inject()(configuration: Configuration) extends ConfigReader {
   object Features {
     val allowOrganizationCreation = get[Boolean]("features.allowOrganizationCreation")
     val enableFrontpage = get[Boolean]("features.enableFrontpage")
+    val isDemoInstance = get[Boolean]("features.isDemoInstance")
   }
 
   val operatorData = get[String]("operatorData")