+++ frontend/javascripts/oxalis/model/data_layer.js
@@ -27,6 +27,7 @@ class DataLayer {
   activeMapping: ?string;
   layerRenderingManager: LayerRenderingManager;
   resolutions: Array<Vector3>;
+  fallbackLayer: ?string;
 
   constructor(
     layerInfo: DataLayerType,
@@ -36,10 +37,12 @@ class DataLayer {
   ) {
     this.connectionInfo = connectionInfo;
     this.name = layerInfo.name;
+    this.fallbackLayer = layerInfo.fallbackLayer != null ? layerInfo.fallbackLayer : null;
     this.resolutions = layerInfo.resolutions;
 
     const { dataset } = Store.getState();
     const bitDepth = getBitDepth(getLayerByName(dataset, this.name));
+    const isSegmentation = layerInfo.category === "segmentation";
 
     ErrorHandling.assert(this.resolutions.length > 0, "Resolutions for layer cannot be empty");
 
@@ -47,7 +50,7 @@ class DataLayer {
       getLayerBoundaries(dataset, this.name).upperBoundary,
       this.resolutions.length,
       bitDepth,
-      layerInfo.category === "segmentation",
+      isSegmentation,
     );
 
     this.pullQueue = new PullQueue(
@@ -58,15 +61,15 @@ class DataLayer {
     );
     this.pushQueue = new PushQueue(this.cube);
     this.cube.initializeWithQueues(this.pullQueue, this.pushQueue);
-    this.mappings = new Mappings(layerInfo.name);
-    this.activeMapping = null;
+    const fallbackLayerName = layerInfo.fallbackLayer != null ? layerInfo.fallbackLayer : null;
+    this.mappings = new Mappings(layerInfo.name, fallbackLayerName);
     this.layerRenderingManager = new LayerRenderingManager(
       this.name,
       this.pullQueue,
       this.cube,
       textureWidth,
       dataTextureCount,
-      layerInfo.category === "segmentation",
+      isSegmentation,
     );
   }