+++ frontend/javascripts/oxalis/view/settings/user_settings_view.js
@@ -9,15 +9,16 @@ import { connect } from "react-redux";
 import React, { PureComponent } from "react";
 import _ from "lodash";
 
+import type { APIDataset } from "admin/api_flow_types";
 import {
-  NumberInputSetting,
-  SwitchSetting,
-  NumberSliderSetting,
-  Vector6InputSetting,
   LogSliderSetting,
+  NumberInputSetting,
+  NumberSliderSetting,
+  SwitchSetting,
+  Vector6InputSetting,
 } from "oxalis/view/settings/setting_input_views";
 import type { UserConfiguration, OxalisState, Tracing } from "oxalis/store";
-import type { APIDataset } from "admin/api_flow_types";
+import { enableMergerMode, disableMergerMode } from "oxalis/merger_mode";
 import {
   enforceSkeletonTracing,
   getActiveNode,
@@ -25,6 +26,7 @@ import {
 import { enforceVolumeTracing } from "oxalis/model/accessors/volumetracing_accessor";
 import { getMaxZoomValue } from "oxalis/model/accessors/flycam_accessor";
 import { getSomeTracing } from "oxalis/model/accessors/tracing_accessor";
+import { hasSegmentation } from "oxalis/model/accessors/dataset_accessor";
 import { setActiveCellAction } from "oxalis/model/actions/volumetracing_actions";
 import {
   setActiveNodeAction,
@@ -36,6 +38,7 @@ import { setUserBoundingBoxAction } from "oxalis/model/actions/annotation_action
 import { setZoomStepAction } from "oxalis/model/actions/flycam_actions";
 import { settings as settingsLabels } from "messages";
 import { updateUserSettingAction } from "oxalis/model/actions/settings_actions";
+import { userSettings } from "libs/user_settings.schema";
 import Constants, {
   type ControlMode,
   ControlModeEnum,
@@ -43,8 +46,7 @@ import Constants, {
   type Vector6,
 } from "oxalis/constants";
 import * as Utils from "libs/utils";
-import { enableMergerMode, disableMergerMode } from "oxalis/merger_mode";
-import { userSettings } from "libs/user_settings.schema";
+
 import MergerModeModalView from "./merger_mode_modal_view";
 
 const { Panel } = Collapse;
@@ -239,10 +241,7 @@ class UserSettingsView extends PureComponent<UserSettingsViewProps, State> {
       const activeNodeRadius = getActiveNode(skeletonTracing)
         .map(activeNode => activeNode.radius)
         .getOrElse(0);
-      const isMergerModeSupported =
-        (this.props.dataset.dataSource.dataLayers || []).find(
-          layer => layer.category === "segmentation" && layer.elementClass === "uint32",
-        ) != null;
+      const isMergerModeSupported = hasSegmentation(this.props.dataset);
       panels.push(
         <Panel header="Nodes & Trees" key="3a">
           <NumberInputSetting
@@ -301,7 +300,7 @@ class UserSettingsView extends PureComponent<UserSettingsViewProps, State> {
             disabled={!isMergerModeSupported}
             tooltipText={
               !isMergerModeSupported
-                ? "The merger mode is only available for datasets with uint32 segmentations."
+                ? "The merger mode is only available for datasets with a segmentation layer."
                 : null
             }
           />